papical_back_end/requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
django-friendship==1.8.1
django-taggit==1.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
